requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
orjson==3.11.3
packaging==25.0
pluggy==1.6.0
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
